requirements.txt
langchain
langgraph
langchain-google-genai
langchain-mcp-adapters
mcp
mlflow
fastapi
uvicorn
python-dotenv
streamlit
